fix: trim recruitment-reminder.service.ts under 300-line limit (ROK-1335)
GitHub CI lint flagged the service file at 315 counted lines (over the
300 max). Fleet's local lint passed earlier because `--fix` happened
to leave the file under the limit there; mode/format drift after the
Codex P2 commit pushed it over on the clean GitHub checkout.

Moved four pure functions out of the service into the helpers module:
- resolveBumpChannel (takes ChannelResolverService as param)
- formatBumpLog
- buildBumpEmbed
- buildBumpRow

No behavior change. 132/132 recruitment-reminder tests still pass,
typecheck clean, lint clean (api workspace).
